requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.21
pydantic>=2.4.2
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
pymysql>=1.1.0
cryptography>=41.0.4
alembic>=1.12.0
bcrypt>=4.3.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
pytz>=2024.1
groq>=0.3.0
mutagen>=1.47.0
email-validator>=2.0.0
requests>=2.31.0
ffmpeg>=1.0
websockets>=12.0
slowapi>=0.1.8
redis>=5.0.1
payos>=0.1.0
